1. 修正TxtRead获取文件读取的位置

+++ base/src/main/java/com/novelbio/base/dataOperate/TxtRead.java
@@ -587,7 +587,7 @@ class TxtRead implements Closeable {
 		if (readByte < 0 || filesize < 0) {
 			return -1;
 		} else {
-			return readByte/filesize;
+			return (double)readByte/filesize;
 		}
 	}
